requirements.txt
a2wsgi==1.10.8
aiobotocore==2.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
anyio==4.5.2
apache-airflow==2.10.4
apache-airflow-providers-amazon==8.27.0
apache-airflow-providers-celery==3.8.2
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.1
apache-airflow-providers-common-sql==1.17.1
apache-airflow-providers-fab==1.4.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.1
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-microsoft-mssql==3.8.0
apache-airflow-providers-postgres==5.11.3
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.6.1
argcomplete==3.5.2
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
boto3==1.35.93
botocore==1.35.93
cachelib==0.9.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
configparser==7.1.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
Deprecated==1.2.15
dill==0.3.1.1
dnspython==2.6.1
email_validator==2.2.0
eventlet==0.38.2
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.5.0
fsspec==2024.12.0
gevent==24.11.1
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.65.5
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
humanize==4.10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
msgspec==0.19.0
multidict==6.1.0
numpy==2.2.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
ply==3.11
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
PyAthena==3.12.2
pycparser==2.22
pydantic==1.10.19
Pygments==2.18.0
PyJWT==2.9.0
pymssql==2.3.2
pyodbc==5.2.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
redshift-connector==2.1.5
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.20.1
s3fs==2024.12.0
s3transfer==0.10.4
scramp==1.4.5
setproctitle==1.3.4
setuptools==75.7.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.45.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==1.26.20
vine==5.1.0
watchtower==3.3.1
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.17.0
WTForms==3.1.2
yarl==1.15.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
